Exemplo movido, testes de new_user_identified
