+++ Controle de Acesso/Servidor Enterprise - C#/App.config
@@ -1,0 +1,6 @@
+<?xml version="1.0"?>
+<configuration>
+    <startup> 
+        
+    <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5"/></startup>
+</configuration>
